Address comments from reviews
