+++ presto-parser/src/main/java/com/facebook/presto/sql/QueryAbridgerUtil.java
@@ -16,7 +16,6 @@ package com.facebook.presto.sql;
 import com.facebook.presto.sql.tree.CallArgument;
 import com.facebook.presto.sql.tree.Expression;
 import com.facebook.presto.sql.tree.Node;
-import com.facebook.presto.sql.tree.OrderBy;
 import com.facebook.presto.sql.tree.Relation;
 import com.facebook.presto.sql.tree.Select;
 import com.facebook.presto.sql.tree.SelectItem;
@@ -25,30 +24,27 @@ import com.facebook.presto.sql.tree.Table;
 import com.facebook.presto.sql.tree.TableSubquery;
 import com.facebook.presto.sql.tree.TransactionMode;
 import com.facebook.presto.sql.tree.Use;
+import com.google.common.collect.ImmutableList;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
-public class AbbreviatorUtil
+public class QueryAbridgerUtil
 {
-    private static List<Class> blackList = new ArrayList<>(Arrays.asList(
+    private static List<Class> blackList = ImmutableList.of(
             Use.class,
-            OrderBy.class,
             Table.class,
-            TableSubquery.class));
+            TableSubquery.class);
 
-    private static List<Class> whiteList = new ArrayList<>(Arrays.asList(
+    private static List<Class> whiteList = ImmutableList.of(
             Expression.class,
-            OrderBy.class,
             Relation.class,
             CallArgument.class,
             TransactionMode.class,
             Select.class,
             Statement.class,
-            SelectItem.class));
+            SelectItem.class);
 
-    private AbbreviatorUtil()
+    private QueryAbridgerUtil()
     {
     }
 
@@ -72,6 +68,35 @@ public class AbbreviatorUtil
         return false;
     }
 
+    /**
+     * A node that is an instance of any of the subclasses of classes mentioned in the whitelist
+     * is allowed to be pruned unless that subclass is present in the blacklist.
+     *
+     * A few examples for clarification:
+     *
+     * - Input: a 'Query' object
+     * - Output: true
+     * - Reason: Query class extends Statement class. Query is not blacklisted. Statement is whitelisted.
+     *
+     * - input: a 'Use' object
+     * - Output: false
+     * - Reason: Use class also extends Statement class. But Use is blacklisted.
+     *
+     * - input: a 'GroupBy' object
+     * - Output: false
+     * - Reason: Groupby is not whitelisted.
+     *
+     * So basically, blacklisted classes are the ones that are
+     * (1) extended from a whitelisted Superclass and
+     * (2) can't be pruned because of their SqlFormatter implementation.
+     *
+     * If we didn't have a blacklist, we would need to whitelist all the allowed subclasses explicitly.
+     * For example, instances of all subclasses of 'Statement', except for 'Use', need to be whitelisted.
+     * We can do that by adding Statement.class to the whitelist and Use.class to blacklist.
+     * If we didn't have a blacklist, we would need to add all the subclasses of Statement (except Use) to the whitelist manually.
+     * @param node
+     * @return boolean
+     */
     public static boolean isAllowedToBePruned(Node node)
     {
         if (isBlackListed(node)) {
